Allow all colour variables to be configured
Only some of the css colour variables were configurable, this adds hugo
config params for every one.

The dot and title colours are also now separately configurable.

Please note: there is also an `--color-base-text` used on `blockquote p`
but it's only read, never set. I chose to ignore it; it might not be
used at all.