Add support for plausible analytics
Plausible is an analytics provider that aims to provide straightforward
analytics that comply with privacy regulations (in particular in the EU,
but other countries are following suit).

Adding support is quite straightforward; all you need is the domain that
you've activated tracking for. This has been added to the Hugo params
and if set the `plausible-analytics.html` partial is included.

It can load alongside google analytics so both partials try to load (and
do nothing if not configured to).

https://plausible.io/
https://plausible.io/docs/plausible-script