Remove date from Tags page
Dates don't make a lot of sense for a tag list page. If a tag can belong
to many posts, made on many days, then what does it mean to put, say, `October
19th 2022` beside tag `foo`?
Is that the first post, last one? Does it matter?

This simply takes the default list template and replaces the `summary`
with a simplified version. I added the 'dot' because I think it looks
nice and breaks the mononoty of an otherwise very bare page.

You could equally take this further and turn it into a full taxonomy
page with sub-content etc. etc. This just takes a first step to making
it usable.