Add canonical link tag
Google Search recommends to point out the canonical link to a page with
a `<link rel="canonical" ...>` tag.

This helps to ensure that Google picks the right page, particularly when
it's faced with http/https pages with the same content. If it's not sure
it can choose not to index the page. This is surfaced in the search
console.

Docs:
 - [Specify a canonical page](https://developers.google.com/search/docs/crawling-indexing/consolidate-duplicate-urls#define-canonical)
 - [<link rel="canonical" ...>](https://developers.google.com/search/docs/crawling-indexing/consolidate-duplicate-urls#rel-canonical-link-method)