Example tags usage
This adds tags to the example site.

The navigation now has a 'tags' link that links to Hugo's default '/tags' page
that lists all tags for a site. Clicking on any tag shows the list of
content with that tag.

I've added tags to the example content. One is deliberately without tags
to make sure that use case is accounted for.

Tags haven't been added to the posts themselves; that can be added, but
needs some thought into where they should go on the page.